+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Segment `buffer-identifiation` now uses face `mode-line-buffer-id`
+  as base.
+
+### Fixed
+
+- Augment `flycheck` now sets variables for `buffer-identification` so
+  that segment running its hooks won't remove the underlining.
+
 ## [v0.7.1]
 
 ### Added